snapcraft.yaml: add dot-kube and home interfaces to kubectl
dot-kube can likely be handled by doing nothing and relying on $HOME
being set to SNAP_USER_DATA. If so, dot-kube can be dropped.

The home interface is not required and users can put files in the
SNAP_USER* directories instead. Having home adds some convenience, but
should possibly request a snap declaration to not auto-connect it (like
it is on classic distro).