add patches/runc-patches/snap-runc-no-pivot-root.patch (#3)
Runc by default uses pivot_root when setting up the container unless
NoPivotRoot is specified in the config (or config doesn't use
configs.NEWNS). When runc is being driven by containerd in the snap and
pivot_root is being used, AppArmor denials are seen with the pivoted
root instead of the pre-pivot path, requiring containerd to have
AppArmor policy with more access than is required. Ideally, the snap
would drive the runc containers with NoPivotRoot, but this is apparently
difficult to standardize in different kubernetes environments (eg, EKS).

snap-runc-no-pivot-root.patch updates runc to continue to use
msMoveRoot() when NoPivotRoot is specified, and to use chroot()
otherwise. Since AppArmor mediates accesses in the chroot with the full
path, containerd no longer needs the expanded access.

References:
https://bugs.launchpad.net/apparmor/+bug/1791711/comments/8